dev: start splitting worker
