+++ pkg/controller/chi/worker.go
@@ -22,7 +22,6 @@ import (
 	"github.com/juliangruber/go-intersect"
 	"gopkg.in/d4l3k/messagediff.v1"
 	core "k8s.io/api/core/v1"
-	v1 "k8s.io/api/policy/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -492,90 +491,6 @@ func (w *worker) logOldAndNew(name string, old, new *chiv1.ClickHouseInstallatio
 	w.logCHI(name+" new", new)
 }
 
-// reconcileCHI run reconcile cycle for a CHI
-func (w *worker) reconcileCHI(ctx context.Context, old, new *chiv1.ClickHouseInstallation) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.logOldAndNew("non-normalized yet", old, new)
-
-	switch {
-	case w.isAfterFinalizerInstalled(old, new):
-		w.a.M(new).F().Info("isAfterFinalizerInstalled - continue reconcile-1")
-	case w.isGenerationTheSame(old, new):
-		w.a.M(new).F().Info("isGenerationTheSame() - nothing to do here, exit")
-		return nil
-	}
-
-	w.a.M(new).S().P()
-	defer w.a.M(new).E().P()
-
-	if new.Status.GetNormalizedCHICompleted() != nil {
-		w.a.M(new).F().Info("has NormalizedCHICompleted, use it as a base for reconcile")
-		old = new.Status.GetNormalizedCHICompleted()
-	}
-
-	old = w.normalize(old)
-	new = w.normalize(new)
-	w.logOldAndNew("normalized", old, new)
-
-	actionPlan := chopmodel.NewActionPlan(old, new)
-	w.logActionPlan(actionPlan)
-
-	switch {
-	case actionPlan.HasActionsToDo():
-		w.a.M(new).F().Info("ActionPlan has actions - continue reconcile")
-	case w.isAfterFinalizerInstalled(old, new):
-		w.a.M(new).F().Info("isAfterFinalizerInstalled - continue reconcile-2")
-	default:
-		w.a.M(new).F().Info("ActionPlan has no actions and not finalizer - nothing to do")
-		return nil
-	}
-
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.newTask(new)
-	w.markReconcileStart(ctx, new, actionPlan)
-	w.excludeStopped(new)
-	w.walkHosts(ctx, new, actionPlan)
-
-	if err := w.reconcile(ctx, new); err != nil {
-		w.a.WithEvent(new, eventActionReconcile, eventReasonReconcileFailed).
-			WithStatusError(new).
-			M(new).F().
-			Error("FAILED update: %v", err)
-		return nil
-	}
-
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	// Post-process added items
-	w.a.V(1).
-		WithEvent(new, eventActionReconcile, eventReasonReconcileInProgress).
-		WithStatusAction(new).
-		M(new).F().
-		Info("remove items scheduled for deletion")
-	w.clear(ctx, new)
-	w.dropReplicas(ctx, new, actionPlan)
-	w.includeStopped(new)
-	w.waitForIPAddresses(ctx, new)
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-	w.markReconcileComplete(ctx, new)
-
-	return nil
-}
-
 func (w *worker) waitForIPAddresses(ctx context.Context, chi *chiv1.ClickHouseInstallation) {
 	if util.IsContextDone(ctx) {
 		log.V(2).Info("task is done")
@@ -620,57 +535,6 @@ func (w *worker) includeStopped(chi *chiv1.ClickHouseInstallation) {
 			Info("add CHI to monitoring")
 		w.c.updateWatch(chi.Namespace, chi.Name, chopmodel.CreateFQDNs(chi, nil, false))
 	}
-}
-
-func (w *worker) clear(ctx context.Context, chi *chiv1.ClickHouseInstallation) {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return
-	}
-
-	// Remove deleted items
-	objs := w.c.discovery(ctx, chi)
-	need := w.task.registryReconciled
-	w.a.V(1).M(chi).F().Info("Reconciled objects:\n%s", w.task.registryReconciled)
-	w.a.V(1).M(chi).F().Info("Existing objects:\n%s", objs)
-	objs.Subtract(need)
-	w.a.V(1).M(chi).F().Info("Non-reconciled objects:\n%s", objs)
-	if w.purge(ctx, chi, objs, w.task.registryFailed) > 0 {
-		w.c.enqueueObject(NewDropDns(&chi.ObjectMeta))
-		util.WaitContextDoneOrTimeout(ctx, 1*time.Minute)
-	}
-
-	w.a.V(1).
-		WithEvent(chi, eventActionReconcile, eventReasonReconcileInProgress).
-		WithStatusAction(chi).
-		M(chi).F().
-		Info("remove items scheduled for deletion")
-}
-
-func (w *worker) dropReplicas(ctx context.Context, chi *chiv1.ClickHouseInstallation, ap *chopmodel.ActionPlan) {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return
-	}
-
-	w.a.V(1).M(chi).F().S().Info("drop replicas based on AP")
-	cnt := 0
-	ap.WalkRemoved(
-		func(cluster *chiv1.ChiCluster) {
-		},
-		func(shard *chiv1.ChiShard) {
-		},
-		func(host *chiv1.ChiHost) {
-			var run *chiv1.ChiHost
-			if shard := host.GetShard(); shard != nil {
-				run = shard.FirstHost()
-			}
-
-			_ = w.dropReplica(ctx, run, host)
-			cnt++
-		},
-	)
-	w.a.V(1).M(chi).F().E().Info("processed replicas: %d", cnt)
 }
 
 func (w *worker) markReconcileStart(ctx context.Context, chi *chiv1.ClickHouseInstallation, ap *chopmodel.ActionPlan) {
@@ -824,129 +688,6 @@ func (w *worker) walkHosts(ctx context.Context, chi *chiv1.ClickHouseInstallatio
 	})
 }
 
-func shouldPurgeStatefulSet(chi *chiv1.ClickHouseInstallation, reconcileFailedObjs *chopmodel.Registry, m meta.ObjectMeta) bool {
-	if reconcileFailedObjs.HasStatefulSet(m) {
-		return chi.GetReconciling().GetCleanup().GetReconcileFailedObjects().GetStatefulSet() == chiv1.ObjectsCleanupDelete
-	}
-	return chi.GetReconciling().GetCleanup().GetUnknownObjects().GetStatefulSet() == chiv1.ObjectsCleanupDelete
-}
-
-func shouldPurgePVC(chi *chiv1.ClickHouseInstallation, reconcileFailedObjs *chopmodel.Registry, m meta.ObjectMeta) bool {
-	if reconcileFailedObjs.HasPVC(m) {
-		return chi.GetReconciling().GetCleanup().GetReconcileFailedObjects().GetPVC() == chiv1.ObjectsCleanupDelete
-	}
-	return chi.GetReconciling().GetCleanup().GetUnknownObjects().GetPVC() == chiv1.ObjectsCleanupDelete
-}
-
-func shouldPurgeConfigMap(chi *chiv1.ClickHouseInstallation, reconcileFailedObjs *chopmodel.Registry, m meta.ObjectMeta) bool {
-	if reconcileFailedObjs.HasConfigMap(m) {
-		return chi.GetReconciling().GetCleanup().GetReconcileFailedObjects().GetConfigMap() == chiv1.ObjectsCleanupDelete
-	}
-	return chi.GetReconciling().GetCleanup().GetUnknownObjects().GetConfigMap() == chiv1.ObjectsCleanupDelete
-}
-
-func shouldPurgeService(chi *chiv1.ClickHouseInstallation, reconcileFailedObjs *chopmodel.Registry, m meta.ObjectMeta) bool {
-	if reconcileFailedObjs.HasService(m) {
-		return chi.GetReconciling().GetCleanup().GetReconcileFailedObjects().GetService() == chiv1.ObjectsCleanupDelete
-	}
-	return chi.GetReconciling().GetCleanup().GetUnknownObjects().GetService() == chiv1.ObjectsCleanupDelete
-}
-
-func shouldPurgeSecret(chi *chiv1.ClickHouseInstallation, reconcileFailedObjs *chopmodel.Registry, m meta.ObjectMeta) bool {
-	if reconcileFailedObjs.HasSecret(m) {
-		return chi.GetReconciling().GetCleanup().GetReconcileFailedObjects().GetSecret() == chiv1.ObjectsCleanupDelete
-	}
-	return chi.GetReconciling().GetCleanup().GetUnknownObjects().GetSecret() == chiv1.ObjectsCleanupDelete
-}
-
-func shouldPurgePDB(chi *chiv1.ClickHouseInstallation, reconcileFailedObjs *chopmodel.Registry, m meta.ObjectMeta) bool {
-	return true
-}
-
-// purge
-func (w *worker) purge(
-	ctx context.Context,
-	chi *chiv1.ClickHouseInstallation,
-	reg *chopmodel.Registry,
-	reconcileFailedObjs *chopmodel.Registry,
-) (cnt int) {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return cnt
-	}
-
-	reg.Walk(func(entityType chopmodel.EntityType, m meta.ObjectMeta) {
-		switch entityType {
-		case chopmodel.StatefulSet:
-			if shouldPurgeStatefulSet(chi, reconcileFailedObjs, m) {
-				w.a.V(1).M(m).F().Info("Delete StatefulSet %s/%s", m.Namespace, m.Name)
-				if err := w.c.kubeClient.AppsV1().StatefulSets(m.Namespace).Delete(ctx, m.Name, newDeleteOptions()); err != nil {
-					w.a.V(1).M(m).F().Error("FAILED to delete StatefulSet %s/%s, err: %v", m.Namespace, m.Name, err)
-				}
-				cnt++
-			}
-		case chopmodel.PVC:
-			if shouldPurgePVC(chi, reconcileFailedObjs, m) {
-				if chopmodel.GetReclaimPolicy(m) == chiv1.PVCReclaimPolicyDelete {
-					w.a.V(1).M(m).F().Info("Delete PVC %s/%s", m.Namespace, m.Name)
-					if err := w.c.kubeClient.CoreV1().PersistentVolumeClaims(m.Namespace).Delete(ctx, m.Name, newDeleteOptions()); err != nil {
-						w.a.V(1).M(m).F().Error("FAILED to delete PVC %s/%s, err: %v", m.Namespace, m.Name, err)
-					}
-				}
-			}
-		case chopmodel.ConfigMap:
-			if shouldPurgeConfigMap(chi, reconcileFailedObjs, m) {
-				w.a.V(1).M(m).F().Info("Delete ConfigMap %s/%s", m.Namespace, m.Name)
-				if err := w.c.kubeClient.CoreV1().ConfigMaps(m.Namespace).Delete(ctx, m.Name, newDeleteOptions()); err != nil {
-					w.a.V(1).M(m).F().Error("FAILED to delete ConfigMap %s/%s, err: %v", m.Namespace, m.Name, err)
-				}
-			}
-		case chopmodel.Service:
-			if shouldPurgeService(chi, reconcileFailedObjs, m) {
-				w.a.V(1).M(m).F().Info("Delete Service %s/%s", m.Namespace, m.Name)
-				if err := w.c.kubeClient.CoreV1().Services(m.Namespace).Delete(ctx, m.Name, newDeleteOptions()); err != nil {
-					w.a.V(1).M(m).F().Error("FAILED to delete Service %s/%s, err: %v", m.Namespace, m.Name, err)
-				}
-			}
-		case chopmodel.Secret:
-			if shouldPurgeSecret(chi, reconcileFailedObjs, m) {
-				w.a.V(1).M(m).F().Info("Delete Secret %s/%s", m.Namespace, m.Name)
-				if err := w.c.kubeClient.CoreV1().Secrets(m.Namespace).Delete(ctx, m.Name, newDeleteOptions()); err != nil {
-					w.a.V(1).M(m).F().Error("FAILED to delete Secret %s/%s, err: %v", m.Namespace, m.Name, err)
-				}
-			}
-		case chopmodel.PDB:
-			if shouldPurgePDB(chi, reconcileFailedObjs, m) {
-				w.a.V(1).M(m).F().Info("Delete PDB %s/%s", m.Namespace, m.Name)
-				if err := w.c.kubeClient.PolicyV1().PodDisruptionBudgets(m.Namespace).Delete(ctx, m.Name, newDeleteOptions()); err != nil {
-					w.a.V(1).M(m).F().Error("FAILED to delete PDB %s/%s, err: %v", m.Namespace, m.Name, err)
-				}
-			}
-		}
-	})
-	return cnt
-}
-
-// reconcile reconciles ClickHouseInstallation
-func (w *worker) reconcile(ctx context.Context, chi *chiv1.ClickHouseInstallation) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(chi).S().P()
-	defer w.a.V(2).M(chi).E().P()
-
-	return chi.WalkTillError(
-		ctx,
-		w.reconcileCHIAuxObjectsPreliminary,
-		w.reconcileCluster,
-		w.reconcileShard,
-		w.reconcileHost,
-		w.reconcileCHIAuxObjectsFinal,
-	)
-}
-
 // baseRemoteServersGeneratorOptions build base set of RemoteServersGeneratorOptions
 // which are applied on each of `remote_servers` reconfiguration during reconcile cycle
 func (w *worker) baseRemoteServersGeneratorOptions() *chopmodel.RemoteServersGeneratorOptions {
@@ -971,135 +712,6 @@ func (w *worker) options(excludeHosts ...*chiv1.ChiHost) *chopmodel.ClickHouseCo
 	return chopmodel.NewClickHouseConfigFilesGeneratorOptions().SetRemoteServersGeneratorOptions(opts)
 }
 
-// reconcileCHIAuxObjectsPreliminary reconciles CHI preliminary in order to ensure that ConfigMaps are in place
-func (w *worker) reconcileCHIAuxObjectsPreliminary(ctx context.Context, chi *chiv1.ClickHouseInstallation) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(chi).S().P()
-	defer w.a.V(2).M(chi).E().P()
-
-	// 1. CHI Service
-	if chi.IsStopped() {
-		// Stopped cluster must have no entry point
-		_ = w.c.deleteServiceCHI(ctx, chi)
-	} else {
-		if service := w.task.creator.CreateServiceCHI(); service != nil {
-			if err := w.reconcileService(ctx, chi, service); err != nil {
-				// Service not reconciled
-				w.task.registryFailed.RegisterService(service.ObjectMeta)
-				return err
-			}
-			w.task.registryReconciled.RegisterService(service.ObjectMeta)
-		}
-	}
-
-	// 2. CHI common ConfigMap without added hosts
-	options := w.options()
-	if err := w.reconcileCHIConfigMapCommon(ctx, chi, options); err != nil {
-		w.a.F().Error("failed to reconcile config map common. err: %v", err)
-	}
-	// 3. CHI users ConfigMap
-	if err := w.reconcileCHIConfigMapUsers(ctx, chi); err != nil {
-		w.a.F().Error("failed to reconcile config map users. err: %v", err)
-	}
-
-	return nil
-}
-
-// reconcileCHIAuxObjectsFinal reconciles CHI global objects
-func (w *worker) reconcileCHIAuxObjectsFinal(ctx context.Context, chi *chiv1.ClickHouseInstallation) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(chi).S().P()
-	defer w.a.V(2).M(chi).E().P()
-
-	// CHI ConfigMaps with update
-	return w.reconcileCHIConfigMapCommon(ctx, chi, nil)
-}
-
-// reconcileCHIConfigMapCommon reconciles all CHI's common ConfigMap
-func (w *worker) reconcileCHIConfigMapCommon(
-	ctx context.Context,
-	chi *chiv1.ClickHouseInstallation,
-	options *chopmodel.ClickHouseConfigFilesGeneratorOptions,
-) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	// ConfigMap common for all resources in CHI
-	// contains several sections, mapped as separated chopConfig files,
-	// such as remote servers, zookeeper setup, etc
-	configMapCommon := w.task.creator.CreateConfigMapCHICommon(options)
-	err := w.reconcileConfigMap(ctx, chi, configMapCommon)
-	if err == nil {
-		w.task.registryReconciled.RegisterConfigMap(configMapCommon.ObjectMeta)
-	} else {
-		w.task.registryFailed.RegisterConfigMap(configMapCommon.ObjectMeta)
-	}
-	return err
-}
-
-// reconcileCHIConfigMapUsers reconciles all CHI's users ConfigMap
-// ConfigMap common for all users resources in CHI
-func (w *worker) reconcileCHIConfigMapUsers(ctx context.Context, chi *chiv1.ClickHouseInstallation) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	// ConfigMap common for all users resources in CHI
-	configMapUsers := w.task.creator.CreateConfigMapCHICommonUsers()
-	err := w.reconcileConfigMap(ctx, chi, configMapUsers)
-	if err == nil {
-		w.task.registryReconciled.RegisterConfigMap(configMapUsers.ObjectMeta)
-	} else {
-		w.task.registryFailed.RegisterConfigMap(configMapUsers.ObjectMeta)
-	}
-	return err
-}
-
-// reconcileHostConfigMap reconciles host's personal ConfigMap
-func (w *worker) reconcileHostConfigMap(ctx context.Context, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	// ConfigMap for a host
-	configMap := w.task.creator.CreateConfigMapHost(host)
-	err := w.reconcileConfigMap(ctx, host.CHI, configMap)
-	if err == nil {
-		w.task.registryReconciled.RegisterConfigMap(configMap.ObjectMeta)
-	} else {
-		w.task.registryFailed.RegisterConfigMap(configMap.ObjectMeta)
-		return err
-	}
-
-	//replicatedObjectNames, replicatedCreateSQLs, distributedObjectNames, distributedCreateSQLs := w.schemer.hostCreateTablesSQLs(task, host)
-	//names := append(replicatedObjectNames, distributedObjectNames...)
-	//sql := append(replicatedCreateSQLs, distributedCreateSQLs...)
-	//
-	// ConfigMap for a host
-	//configMap = w.creator.CreateConfigMapHostMigration(host, w.creator.MakeConfigMapData(names, sql))
-	//err = w.reconcileConfigMap(task, host.CHI, configMap)
-	//if err == nil {
-	//	w.registryReconciled.RegisterConfigMap(configMap.ObjectMeta)
-	//} else {
-	//	w.registryFailed.RegisterConfigMap(configMap.ObjectMeta)
-	//	return err
-	//}
-
-	return nil
-}
-
 // prepareHostStatefulSetWithStatus prepares host's StatefulSet status
 func (w *worker) prepareHostStatefulSetWithStatus(ctx context.Context, host *chiv1.ChiHost, shutdown bool) {
 	if util.IsContextDone(ctx) {
@@ -1110,174 +722,6 @@ func (w *worker) prepareHostStatefulSetWithStatus(ctx context.Context, host *chi
 	// StatefulSet for a host
 	_ = w.task.creator.CreateStatefulSet(host, shutdown)
 	(&host.ReconcileAttributes).SetStatus(w.getStatefulSetStatus(host.StatefulSet.ObjectMeta))
-}
-
-// reconcileHostStatefulSet reconciles host's StatefulSet
-func (w *worker) reconcileHostStatefulSet(ctx context.Context, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	// In case we have to force-restart host
-	// We'll do it via replicas: 0 in StatefulSet.
-	if w.shouldForceRestartHost(host) {
-		w.prepareHostStatefulSetWithStatus(ctx, host, true)
-		_ = w.reconcileStatefulSet(ctx, host)
-		// At this moment StatefulSet has 0 replicas.
-		// First stage of RollingUpdate completed.
-	}
-
-	// We are in place, where we can  reconcile StatefulSet to desired configuration.
-	w.prepareHostStatefulSetWithStatus(ctx, host, false)
-	err := w.reconcileStatefulSet(ctx, host)
-	if err == nil {
-		w.task.registryReconciled.RegisterStatefulSet(host.StatefulSet.ObjectMeta)
-	} else {
-		w.task.registryFailed.RegisterStatefulSet(host.StatefulSet.ObjectMeta)
-		if err == errIgnore {
-			err = nil
-		}
-	}
-	return err
-}
-
-// reconcileHostService reconciles host's Service
-func (w *worker) reconcileHostService(ctx context.Context, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-	service := w.task.creator.CreateServiceHost(host)
-	if service == nil {
-		// This is not a problem, service may be omitted
-		return nil
-	}
-	err := w.reconcileService(ctx, host.CHI, service)
-	if err == nil {
-		w.task.registryReconciled.RegisterService(service.ObjectMeta)
-	} else {
-		w.task.registryFailed.RegisterService(service.ObjectMeta)
-	}
-	return err
-}
-
-// reconcileCluster reconciles Cluster, excluding nested shards
-func (w *worker) reconcileCluster(ctx context.Context, cluster *chiv1.ChiCluster) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(cluster).S().P()
-	defer w.a.V(2).M(cluster).E().P()
-
-	// Add Cluster's Service
-	if service := w.task.creator.CreateServiceCluster(cluster); service != nil {
-		if err := w.reconcileService(ctx, cluster.CHI, service); err == nil {
-			w.task.registryReconciled.RegisterService(service.ObjectMeta)
-		} else {
-			w.task.registryFailed.RegisterService(service.ObjectMeta)
-		}
-	}
-
-	// Add Cluster's Auto Secret
-	if cluster.Secret.Source() == chiv1.ClusterSecretSourceAuto {
-		if secret := w.task.creator.CreateClusterSecret(chopmodel.CreateClusterAutoSecretName(cluster)); secret != nil {
-			if err := w.reconcileSecret(ctx, cluster.CHI, secret); err == nil {
-				w.task.registryReconciled.RegisterSecret(secret.ObjectMeta)
-			} else {
-				w.task.registryFailed.RegisterSecret(secret.ObjectMeta)
-			}
-		}
-	}
-
-	pdb := w.task.creator.NewPodDisruptionBudget(cluster)
-	if err := w.reconcilePDB(ctx, cluster, pdb); err == nil {
-		w.task.registryReconciled.RegisterPDB(pdb.ObjectMeta)
-	} else {
-		w.task.registryFailed.RegisterPDB(pdb.ObjectMeta)
-	}
-
-	return nil
-}
-
-// reconcileShard reconciles Shard, excluding nested replicas
-func (w *worker) reconcileShard(ctx context.Context, shard *chiv1.ChiShard) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(shard).S().P()
-	defer w.a.V(2).M(shard).E().P()
-
-	// Add Shard's Service
-	service := w.task.creator.CreateServiceShard(shard)
-	if service == nil {
-		// This is not a problem, ServiceShard may be omitted
-		return nil
-	}
-	err := w.reconcileService(ctx, shard.CHI, service)
-	if err == nil {
-		w.task.registryReconciled.RegisterService(service.ObjectMeta)
-	} else {
-		w.task.registryFailed.RegisterService(service.ObjectMeta)
-	}
-	return err
-}
-
-// reconcileHost reconciles ClickHouse host
-func (w *worker) reconcileHost(ctx context.Context, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(host).S().P()
-	defer w.a.V(2).M(host).E().P()
-
-	w.a.V(1).
-		WithEvent(host.GetCHI(), eventActionReconcile, eventReasonReconcileStarted).
-		WithStatusAction(host.GetCHI()).
-		M(host).F().
-		Info("Reconcile Host %s started", host.Name)
-
-	// Create artifacts
-	w.prepareHostStatefulSetWithStatus(ctx, host, false)
-
-	if err := w.excludeHost(ctx, host); err != nil {
-		return err
-	}
-
-	if err := w.reconcileHostConfigMap(ctx, host); err != nil {
-		return err
-	}
-	//w.reconcilePersistentVolumes(task, host)
-	_ = w.reconcilePVCs(ctx, host)
-	if err := w.reconcileHostStatefulSet(ctx, host); err != nil {
-		return err
-	}
-	_ = w.reconcilePVCs(ctx, host)
-
-	_ = w.reconcileHostService(ctx, host)
-
-	host.ReconcileAttributes.UnsetAdd()
-
-	_ = w.migrateTables(ctx, host)
-
-	if err := w.includeHost(ctx, host); err != nil {
-		// If host is not ready - fallback
-		return err
-	}
-
-	w.a.V(1).
-		WithEvent(host.CHI, eventActionReconcile, eventReasonReconcileCompleted).
-		WithStatusAction(host.CHI).
-		M(host).F().
-		Info("Reconcile Host %s completed", host.Name)
-
-	return nil
 }
 
 // migrateTables
@@ -1534,398 +978,6 @@ func (w *worker) waitHostNoActiveQueries(ctx context.Context, host *chiv1.ChiHos
 	})
 }
 
-// reconcilePDB reconciles PodDisruptionBudget
-func (w *worker) reconcilePDB(ctx context.Context, cluster *chiv1.ChiCluster, pdb *v1.PodDisruptionBudget) error {
-	_, err := w.c.kubeClient.PolicyV1().PodDisruptionBudgets(pdb.Namespace).Get(ctx, pdb.Name, newGetOptions())
-	switch {
-	case err == nil:
-		_, err := w.c.kubeClient.PolicyV1().PodDisruptionBudgets(pdb.Namespace).Update(ctx, pdb, newUpdateOptions())
-		if err == nil {
-			log.V(1).Info("PDB updated %s/%s", pdb.Namespace, pdb.Name)
-		}
-	case (err != nil) && apierrors.IsNotFound(err):
-		_, err := w.c.kubeClient.PolicyV1().PodDisruptionBudgets(pdb.Namespace).Create(ctx, pdb, newCreateOptions())
-		if err == nil {
-			log.V(1).Info("PDB created %s/%s", pdb.Namespace, pdb.Name)
-		}
-	}
-	return nil
-}
-
-// deleteCHI
-func (w *worker) deleteCHI(ctx context.Context, old, new *chiv1.ClickHouseInstallation) bool {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return false
-	}
-
-	// Do we have pending request for CHI to be deleted?
-	if new.ObjectMeta.DeletionTimestamp.IsZero() {
-		// CHI is not being deleted and operator has not deleted anything.
-		return false
-	}
-
-	w.a.V(3).M(new).S().P()
-	defer w.a.V(3).M(new).E().P()
-
-	// Ok, we have pending request for CHI to be deleted.
-	// However, we need to decide, should CHI's child resources be deleted or not.
-	// There is a curious situation, when CRD is deleted and k8s starts to delete all resources of the type,
-	// described by CRD being deleted. This is may be unexpected and very painful situation,
-	// so in this case we should agree to delete CHI itself, but has to keep all CHI's child resources.
-
-	var clear bool
-	crd, err := w.c.extClient.ApiextensionsV1().CustomResourceDefinitions().Get(ctx, "clickhouseinstallations.clickhouse.altinity.com", newGetOptions())
-	if err == nil {
-		if crd.ObjectMeta.DeletionTimestamp.IsZero() {
-			// CRD is not being deleted and operator can delete all child resources.
-			w.a.V(1).M(new).F().Info("CRD %s/%s is not being deleted, operator will delete child resources", crd.Namespace, crd.Name)
-			clear = true
-		} else {
-			// CRD is being deleted. This may be a mistake, operator should not delete data
-			w.a.V(1).M(new).F().Info("CRD %s/%s BEING DELETED, operator will NOT delete child resources", crd.Namespace, crd.Name)
-			clear = false
-		}
-	} else {
-		w.a.V(1).M(new).F().Error("unable to get CRD, got error: %v ", err)
-		w.a.V(1).M(new).F().Info("will delete chi %s/%s", new.Namespace, new.Name)
-		clear = true
-	}
-
-	if clear {
-		cur, err := w.c.chopClient.ClickhouseV1().ClickHouseInstallations(new.Namespace).Get(ctx, new.Name, newGetOptions())
-		if cur == nil {
-			return false
-		}
-		if err != nil {
-			return false
-		}
-
-		if !util.InArray(FinalizerName, new.ObjectMeta.Finalizers) {
-			// No finalizer found, unexpected behavior
-			return false
-		}
-
-		_ = w.deleteCHIProtocol(ctx, new)
-	} else {
-		new.Attributes.SkipOwnerRef = true
-		_ = w.reconcileCHI(ctx, old, new)
-	}
-
-	// We need to uninstall finalizer in order to allow k8s to delete CHI resource
-	w.a.V(2).M(new).F().Info("uninstall finalizer")
-	if err := w.c.uninstallFinalizer(ctx, new); err != nil {
-		w.a.V(1).M(new).F().Error("unable to uninstall finalizer: err:%v", err)
-	}
-
-	// CHI's child resources were deleted
-	return true
-}
-
-// discoveryAndDeleteCHI deletes all kubernetes resources related to chi *chop.ClickHouseInstallation
-func (w *worker) discoveryAndDeleteCHI(ctx context.Context, chi *chiv1.ClickHouseInstallation) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	objs := w.c.discovery(ctx, chi)
-	if objs.NumStatefulSet() > 0 {
-		chi.WalkHosts(func(host *chiv1.ChiHost) error {
-			_ = w.schemer.HostSyncTables(ctx, host)
-			return nil
-		})
-	}
-	w.purge(ctx, chi, objs, nil)
-	return nil
-}
-
-// deleteCHIProtocol deletes all kubernetes resources related to chi *chop.ClickHouseInstallation
-func (w *worker) deleteCHIProtocol(ctx context.Context, chi *chiv1.ClickHouseInstallation) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(chi).S().P()
-	defer w.a.V(2).M(chi).E().P()
-
-	var err error
-	chi, err = w.normalizer.CreateTemplatedCHI(chi, chopmodel.NewNormalizerOptions())
-	if err != nil {
-		w.a.WithEvent(chi, eventActionDelete, eventReasonDeleteFailed).
-			WithStatusError(chi).
-			M(chi).F().
-			Error("Delete CHI failed - unable to normalize: %q", err)
-		return err
-	}
-
-	// Announce delete procedure
-	w.a.V(1).
-		WithEvent(chi, eventActionDelete, eventReasonDeleteStarted).
-		WithStatusAction(chi).
-		M(chi).F().
-		Info("Delete CHI started")
-
-	chi.EnsureStatus().DeleteStart()
-	if err := w.c.updateCHIObjectStatus(ctx, chi, UpdateCHIStatusOptions{
-		TolerateAbsence: true,
-		CopyCHIStatusOptions: chiv1.CopyCHIStatusOptions{
-			MainFields: true,
-		},
-	}); err != nil {
-		w.a.V(1).M(chi).F().Error("UNABLE to write normalized CHI. err:%q", err)
-		return nil
-	}
-
-	// Start delete protocol
-
-	// Exclude this CHI from monitoring
-	w.c.deleteWatch(chi.Namespace, chi.Name)
-
-	// Delete Service
-	_ = w.c.deleteServiceCHI(ctx, chi)
-
-	chi.WalkHosts(func(host *chiv1.ChiHost) error {
-		_ = w.schemer.HostSyncTables(ctx, host)
-		return nil
-	})
-
-	// Delete all clusters
-	chi.WalkClusters(func(cluster *chiv1.ChiCluster) error {
-		return w.deleteCluster(ctx, chi, cluster)
-	})
-
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	// Delete ConfigMap(s)
-	_ = w.c.deleteConfigMapsCHI(ctx, chi)
-
-	w.a.V(1).
-		WithEvent(chi, eventActionDelete, eventReasonDeleteCompleted).
-		WithStatusAction(chi).
-		M(chi).F().
-		Info("Delete CHI completed")
-
-	return nil
-}
-
-// canDropReplica
-func (w *worker) canDropReplica(host *chiv1.ChiHost) (can bool) {
-	can = true
-	w.c.walkDiscoveredPVCs(host, func(pvc *core.PersistentVolumeClaim) {
-		// Replica's state has to be kept in Zookeeper for retained volumes.
-		// ClickHouse expects to have state of the non-empty replica in-place when replica rejoins.
-		if chopmodel.GetReclaimPolicy(pvc.ObjectMeta) == chiv1.PVCReclaimPolicyRetain {
-			w.a.V(1).F().Info("PVC %s/%s blocks drop replica. Reclaim policy: %s", chiv1.PVCReclaimPolicyRetain.String())
-			can = false
-		}
-	})
-	return can
-}
-
-// dropReplica
-func (w *worker) dropReplica(ctx context.Context, hostToRun, hostToDrop *chiv1.ChiHost) error {
-	if (hostToRun == nil) || (hostToDrop == nil) {
-		w.a.V(1).F().Error("FAILED to drop replica. hostToRun:%s, hostToDrop:%s", hostToRun.GetName(), hostToDrop.GetName())
-		return nil
-	}
-
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-	if !w.canDropReplica(hostToDrop) {
-		w.a.V(1).F().Warning("UNABLE to drop replica. hostToRun:%s, hostToDrop:%s", hostToRun.GetName(), hostToDrop.GetName())
-		return nil
-	}
-
-	err := w.schemer.HostDropReplica(ctx, hostToRun, hostToDrop)
-
-	if err == nil {
-		w.a.V(1).
-			WithEvent(hostToRun.CHI, eventActionDelete, eventReasonDeleteCompleted).
-			WithStatusAction(hostToRun.CHI).
-			M(hostToRun).F().
-			Info("Drop replica host %s in cluster %s", hostToDrop.Name, hostToDrop.Address.ClusterName)
-	} else {
-		w.a.WithEvent(hostToRun.CHI, eventActionDelete, eventReasonDeleteFailed).
-			WithStatusError(hostToRun.CHI).
-			M(hostToRun).F().
-			Error("FAILED to drop replica on host %s with error %v", hostToDrop.Name, err)
-	}
-
-	return err
-}
-
-// deleteTables
-func (w *worker) deleteTables(ctx context.Context, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	if !chopmodel.HostCanDeleteAllPVCs(host) {
-		return nil
-	}
-	err := w.schemer.HostDropTables(ctx, host)
-
-	if err == nil {
-		w.a.V(1).
-			WithEvent(host.CHI, eventActionDelete, eventReasonDeleteCompleted).
-			WithStatusAction(host.CHI).
-			M(host).F().
-			Info("Deleted tables on host %s replica %d to shard %d in cluster %s",
-				host.Name, host.Address.ReplicaIndex, host.Address.ShardIndex, host.Address.ClusterName)
-	} else {
-		w.a.WithEvent(host.CHI, eventActionDelete, eventReasonDeleteFailed).
-			WithStatusError(host.CHI).
-			M(host).F().
-			Error("FAILED to delete tables on host %s with error %v", host.Name, err)
-	}
-
-	return err
-}
-
-// deleteHost deletes all kubernetes resources related to host
-// chi is the new CHI in which there will be no more this host
-func (w *worker) deleteHost(ctx context.Context, chi *chiv1.ClickHouseInstallation, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(host).S().Info(host.Address.HostName)
-	defer w.a.V(2).M(host).E().Info(host.Address.HostName)
-
-	w.a.V(1).
-		WithEvent(host.CHI, eventActionDelete, eventReasonDeleteStarted).
-		WithStatusAction(host.CHI).
-		M(host).F().
-		Info("Delete host %s/%s - started", host.Address.ClusterName, host.Name)
-
-	if _, err := w.c.getStatefulSet(host); err != nil {
-		w.a.WithEvent(host.CHI, eventActionDelete, eventReasonDeleteCompleted).
-			WithStatusAction(host.CHI).
-			M(host).F().
-			Info("Delete host %s/%s - completed StatefulSet not found - already deleted? err: %v",
-				host.Address.ClusterName, host.Name, err)
-		return nil
-	}
-
-	// Each host consists of
-	// 1. User-level objects - tables on the host
-	//    We need to delete tables on the host in order to clean Zookeeper data.
-	//    If just delete tables, Zookeeper will still keep track of non-existent tables
-	// 2. Kubernetes-level objects - such as StatefulSet, PVC(s), ConfigMap(s), Service(s)
-	// Need to delete all these items
-
-	var err error
-	_ = w.deleteTables(ctx, host)
-	err = w.c.deleteHost(ctx, host)
-
-	// When deleting the whole CHI (not particular host), CHI may already be unavailable, so update CHI tolerantly
-	chi.EnsureStatus().DeletedHostsCount++
-	_ = w.c.updateCHIObjectStatus(ctx, chi, UpdateCHIStatusOptions{
-		TolerateAbsence: true,
-		CopyCHIStatusOptions: chiv1.CopyCHIStatusOptions{
-			MainFields: true,
-		},
-	})
-
-	if err == nil {
-		w.a.V(1).
-			WithEvent(host.CHI, eventActionDelete, eventReasonDeleteCompleted).
-			WithStatusAction(host.CHI).
-			M(host).F().
-			Info("Delete host %s/%s - completed", host.Address.ClusterName, host.Name)
-	} else {
-		w.a.WithEvent(host.CHI, eventActionDelete, eventReasonDeleteFailed).
-			WithStatusError(host.CHI).
-			M(host).F().
-			Error("FAILED Delete host %s/%s - completed", host.Address.ClusterName, host.Name)
-	}
-
-	return err
-}
-
-// deleteShard deletes all kubernetes resources related to shard *chop.ChiShard
-// chi is the new CHI in which there will be no more this shard
-func (w *worker) deleteShard(ctx context.Context, chi *chiv1.ClickHouseInstallation, shard *chiv1.ChiShard) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(shard).S().P()
-	defer w.a.V(2).M(shard).E().P()
-
-	w.a.V(1).
-		WithEvent(shard.CHI, eventActionDelete, eventReasonDeleteStarted).
-		WithStatusAction(shard.CHI).
-		M(shard).F().
-		Info("Delete shard %s/%s - started", shard.Address.Namespace, shard.Name)
-
-	// Delete Shard Service
-	_ = w.c.deleteServiceShard(ctx, shard)
-
-	// Delete all replicas
-	shard.WalkHosts(func(host *chiv1.ChiHost) error {
-		return w.deleteHost(ctx, chi, host)
-	})
-
-	w.a.V(1).
-		WithEvent(shard.CHI, eventActionDelete, eventReasonDeleteCompleted).
-		WithStatusAction(shard.CHI).
-		M(shard).F().
-		Info("Delete shard %s/%s - completed", shard.Address.Namespace, shard.Name)
-
-	return nil
-}
-
-// deleteCluster deletes all kubernetes resources related to cluster *chop.ChiCluster
-// chi is the new CHI in which there will be no more this cluster
-func (w *worker) deleteCluster(ctx context.Context, chi *chiv1.ClickHouseInstallation, cluster *chiv1.ChiCluster) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(cluster).S().P()
-	defer w.a.V(2).M(cluster).E().P()
-
-	w.a.V(1).
-		WithEvent(cluster.CHI, eventActionDelete, eventReasonDeleteStarted).
-		WithStatusAction(cluster.CHI).
-		M(cluster).F().
-		Info("Delete cluster %s/%s - started", cluster.Address.Namespace, cluster.Name)
-
-	// Delete Cluster Service
-	_ = w.c.deleteServiceCluster(ctx, cluster)
-
-	// Delete Cluster's Auto Secret
-	if cluster.Secret.Source() == chiv1.ClusterSecretSourceAuto {
-		// Delete Cluster Secret
-		_ = w.c.deleteSecretCluster(ctx, cluster)
-	}
-
-	// Delete all shards
-	cluster.WalkShards(func(index int, shard *chiv1.ChiShard) error {
-		return w.deleteShard(ctx, chi, shard)
-	})
-
-	w.a.V(1).
-		WithEvent(cluster.CHI, eventActionDelete, eventReasonDeleteCompleted).
-		WithStatusAction(cluster.CHI).
-		M(cluster).F().
-		Info("Delete cluster %s/%s - completed", cluster.Address.Namespace, cluster.Name)
-
-	return nil
-}
-
 // createCHIFromObjectMeta
 func (w *worker) createCHIFromObjectMeta(objectMeta *meta.ObjectMeta, isCHI bool, options *chopmodel.NormalizerOptions) (*chiv1.ClickHouseInstallation, error) {
 	w.a.V(3).M(objectMeta).S().P()
@@ -1992,44 +1044,6 @@ func (w *worker) createConfigMap(ctx context.Context, chi *chiv1.ClickHouseInsta
 			WithStatusError(chi).
 			M(chi).F().
 			Error("Create ConfigMap %s/%s failed with error %v", configMap.Namespace, configMap.Name, err)
-	}
-
-	return err
-}
-
-// reconcileConfigMap reconciles core.ConfigMap which belongs to specified CHI
-func (w *worker) reconcileConfigMap(
-	ctx context.Context,
-	chi *chiv1.ClickHouseInstallation,
-	configMap *core.ConfigMap,
-) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(chi).S().P()
-	defer w.a.V(2).M(chi).E().P()
-
-	// Check whether this object already exists in k8s
-	curConfigMap, err := w.c.getConfigMap(&configMap.ObjectMeta, true)
-
-	if curConfigMap != nil {
-		// We have ConfigMap - try to update it
-		err = w.updateConfigMap(ctx, chi, configMap)
-	}
-
-	if apierrors.IsNotFound(err) {
-		// ConfigMap not found - even during Update process - try to create it
-		err = w.createConfigMap(ctx, chi, configMap)
-	}
-
-	if err != nil {
-		w.a.WithEvent(chi, eventActionReconcile, eventReasonReconcileFailed).
-			WithStatusAction(chi).
-			WithStatusError(chi).
-			M(chi).F().
-			Error("FAILED to reconcile ConfigMap: %s CHI: %s ", configMap.Name, chi.Name)
 	}
 
 	return err
@@ -2144,41 +1158,6 @@ func (w *worker) createService(ctx context.Context, chi *chiv1.ClickHouseInstall
 	return err
 }
 
-// reconcileService reconciles core.Service
-func (w *worker) reconcileService(ctx context.Context, chi *chiv1.ClickHouseInstallation, service *core.Service) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(chi).S().Info(service.Name)
-	defer w.a.V(2).M(chi).E().Info(service.Name)
-
-	// Check whether this object already exists
-	curService, err := w.c.getService(service)
-
-	if curService != nil {
-		// We have Service - try to update it
-		err = w.updateService(ctx, chi, curService, service)
-	}
-
-	if err != nil {
-		// Service not found or not updated. Try to recreate
-		_ = w.c.deleteServiceIfExists(ctx, service.Namespace, service.Name)
-		err = w.createService(ctx, chi, service)
-	}
-
-	if err != nil {
-		w.a.WithEvent(chi, eventActionReconcile, eventReasonReconcileFailed).
-			WithStatusAction(chi).
-			WithStatusError(chi).
-			M(chi).F().
-			Error("FAILED to reconcile Service: %s CHI: %s ", service.Name, chi.Name)
-	}
-
-	return err
-}
-
 // createSecret
 func (w *worker) createSecret(ctx context.Context, chi *chiv1.ClickHouseInstallation, secret *core.Secret) error {
 	if util.IsContextDone(ctx) {
@@ -2199,36 +1178,6 @@ func (w *worker) createSecret(ctx context.Context, chi *chiv1.ClickHouseInstalla
 			WithStatusError(chi).
 			M(chi).F().
 			Error("Create Secret %s/%s failed with error %v", secret.Namespace, secret.Name, err)
-	}
-
-	return err
-}
-
-// reconcileSecret reconciles core.Secret
-func (w *worker) reconcileSecret(ctx context.Context, chi *chiv1.ClickHouseInstallation, secret *core.Secret) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	w.a.V(2).M(chi).S().Info(secret.Name)
-	defer w.a.V(2).M(chi).E().Info(secret.Name)
-
-	// Check whether this object already exists
-	if _, err := w.c.getSecret(secret); err == nil {
-		// We have Secret - try to update it
-		return nil
-	}
-
-	// Secret not found or broken. Try to recreate
-	_ = w.c.deleteSecretIfExists(ctx, secret.Namespace, secret.Name)
-	err := w.createSecret(ctx, chi, secret)
-	if err != nil {
-		w.a.WithEvent(chi, eventActionReconcile, eventReasonReconcileFailed).
-			WithStatusAction(chi).
-			WithStatusError(chi).
-			M(chi).F().
-			Error("FAILED to reconcile Secret: %s CHI: %s ", secret.Name, chi.Name)
 	}
 
 	return err
@@ -2278,51 +1227,6 @@ func (w *worker) getStatefulSetStatus(meta meta.ObjectMeta) chiv1.StatefulSetSta
 	}
 
 	return chiv1.StatefulSetStatusUnknown
-}
-
-// reconcileStatefulSet reconciles apps.StatefulSet
-func (w *worker) reconcileStatefulSet(ctx context.Context, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil
-	}
-
-	newStatefulSet := host.StatefulSet
-
-	w.a.V(2).M(host).S().Info(util.NamespaceNameString(newStatefulSet.ObjectMeta))
-	defer w.a.V(2).M(host).E().Info(util.NamespaceNameString(newStatefulSet.ObjectMeta))
-
-	if host.ReconcileAttributes.GetStatus() == chiv1.StatefulSetStatusSame {
-		defer w.a.V(2).M(host).F().Info("no need to reconcile the same StatefulSet %s", util.NamespaceNameString(newStatefulSet.ObjectMeta))
-		return nil
-	}
-
-	// Check whether this object already exists in k8s
-	var err error
-	host.CurStatefulSet, err = w.c.getStatefulSet(&newStatefulSet.ObjectMeta, false)
-
-	if host.CurStatefulSet != nil {
-		// We have StatefulSet - try to update it
-		err = w.updateStatefulSet(ctx, host)
-	}
-
-	if apierrors.IsNotFound(err) {
-		// StatefulSet not found - even during Update process - try to create it
-		err = w.createStatefulSet(ctx, host)
-	}
-
-	if err != nil {
-		w.a.WithEvent(host.CHI, eventActionReconcile, eventReasonReconcileFailed).
-			WithStatusAction(host.CHI).
-			WithStatusError(host.CHI).
-			M(host).F().
-			Error("FAILED to reconcile StatefulSet: %s CHI: %s ", newStatefulSet.Name, host.CHI.Name)
-	}
-
-	// Host has to know current StatefulSet and Pod
-	host.CurStatefulSet, _ = w.c.getStatefulSet(&newStatefulSet.ObjectMeta, false)
-
-	return err
 }
 
 // createStatefulSet
@@ -2487,91 +1391,6 @@ func (w *worker) recreateStatefulSet(ctx context.Context, host *chiv1.ChiHost) e
 	_ = w.reconcilePVCs(ctx, host)
 	host.StatefulSet = host.DesiredStatefulSet
 	return w.createStatefulSet(ctx, host)
-}
-
-// reconcilePersistentVolumes
-func (w *worker) reconcilePersistentVolumes(ctx context.Context, host *chiv1.ChiHost) {
-	if util.IsContextDone(ctx) {
-		return
-	}
-
-	w.c.walkPVs(host, func(pv *core.PersistentVolume) {
-		pv = w.task.creator.PreparePersistentVolume(pv, host)
-		_, _ = w.c.updatePersistentVolume(ctx, pv)
-	})
-}
-
-// reconcilePVCs
-func (w *worker) reconcilePVCs(ctx context.Context, host *chiv1.ChiHost) error {
-	if util.IsContextDone(ctx) {
-		return nil
-	}
-
-	namespace := host.Address.Namespace
-	w.a.V(2).M(host).S().Info("host %s/%s", namespace, host.Name)
-	defer w.a.V(2).M(host).E().Info("host %s/%s", namespace, host.Name)
-
-	host.WalkVolumeMounts(func(volumeMount *core.VolumeMount) {
-		if util.IsContextDone(ctx) {
-			return
-		}
-
-		volumeClaimTemplateName := volumeMount.Name
-		volumeClaimTemplate, ok := host.CHI.GetVolumeClaimTemplate(volumeClaimTemplateName)
-		if !ok {
-			// No this is not a reference to VolumeClaimTemplate, it may be reference to ConfigMap
-			return
-		}
-
-		pvcName := chopmodel.CreatePVCName(host, volumeMount, volumeClaimTemplate)
-		w.a.V(2).M(host).Info("reconcile volumeMount (%s/%s/%s/%s) - start", namespace, host.Name, volumeMount.Name, pvcName)
-		defer w.a.V(2).M(host).Info("reconcile volumeMount (%s/%s/%s/%s) - end", namespace, host.Name, volumeMount.Name, pvcName)
-
-		pvc, err := w.c.kubeClient.CoreV1().PersistentVolumeClaims(namespace).Get(ctx, pvcName, newGetOptions())
-		if err != nil {
-			if apierrors.IsNotFound(err) {
-				// This is not an error per se, means PVC is not created (yet)?
-				if w.task.creator.OperatorShouldCreatePVC(host, volumeClaimTemplate) {
-					pvc = w.task.creator.CreatePVC(pvcName, host, &volumeClaimTemplate.Spec)
-				} else {
-					// Not created and we are not expected to create PVC by ourselves
-					return
-				}
-			} else {
-				// Any non-NotFound API error - unable to proceed
-				w.a.M(host).F().Error("ERROR unable to get PVC(%s/%s) err: %v", namespace, pvcName, err)
-				return
-			}
-		}
-
-		pvc, err = w.reconcilePVC(ctx, pvc, host, volumeClaimTemplate)
-		if err != nil {
-			w.a.M(host).F().Error("ERROR unable to reconcile PVC(%s/%s) err: %v", namespace, pvcName, err)
-			w.task.registryFailed.RegisterPVC(pvc.ObjectMeta)
-			return
-		}
-
-		w.task.registryReconciled.RegisterPVC(pvc.ObjectMeta)
-	})
-
-	return nil
-}
-
-// reconcilePVC
-func (w *worker) reconcilePVC(
-	ctx context.Context,
-	pvc *core.PersistentVolumeClaim,
-	host *chiv1.ChiHost,
-	template *chiv1.ChiVolumeClaimTemplate,
-) (*core.PersistentVolumeClaim, error) {
-	if util.IsContextDone(ctx) {
-		log.V(2).Info("task is done")
-		return nil, fmt.Errorf("task is done")
-	}
-
-	w.applyPVCResourcesRequests(pvc, template)
-	pvc = w.task.creator.PreparePersistentVolumeClaim(pvc, host, template)
-	return w.c.updatePersistentVolumeClaim(ctx, pvc)
 }
 
 // applyPVCResourcesRequests
